Created s3select package for presto-hive Refactored S3Select related classes in own package
