+++ presto-hive/src/main/java/com/facebook/presto/hive/s3select/S3SelectRecordCursorProvider.java
@@ -11,10 +11,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.s3select;
 
 import com.facebook.presto.common.predicate.TupleDomain;
 import com.facebook.presto.common.type.TypeManager;
+import com.facebook.presto.hive.HdfsEnvironment;
+import com.facebook.presto.hive.HiveClientConfig;
+import com.facebook.presto.hive.HiveColumnHandle;
+import com.facebook.presto.hive.HiveFileSplit;
+import com.facebook.presto.hive.HiveRecordCursorProvider;
+import com.facebook.presto.hive.IonSqlQueryBuilder;
 import com.facebook.presto.hive.s3.PrestoS3ClientFactory;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.PrestoException;
