+++ presto-hive/src/test/java/com/facebook/presto/hive/s3select/TestS3SelectLineRecordReader.java
@@ -11,13 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.s3select;
 
+import com.facebook.presto.hive.s3select.S3SelectLineRecordReader.UnrecoverableS3OperationException;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-
-import static com.facebook.presto.hive.S3SelectLineRecordReader.UnrecoverableS3OperationException;
 
 public class TestS3SelectLineRecordReader
 {
